Refactor camera and shader; add controls
Switch numeric types to float and glm::vec3 across BlackHole and Constants; remove legacy Ray and camera.cpp. Replace Camera with a spherical-parameter implementation (radius/azimuth/polar), expose position and view matrix, and add mutators. Add Shader::setMat4. Update fragment shader to use camPos, invView, aspect ratio, higher step count and finer dt, and adjust ray direction and base color. Update Display to send invView, camPos and aspect ratio uniforms. Main now constructs the new Camera, registers mouse/scroll callbacks for interactive orbit/zoom, toggles pause with space, and prints camera/debug info. Overall: enable single-precision math and interactive camera controls, and fix ray orientation using the inverse view.

+++ BlackHoleTracer/Headers/blackhole.h
@@ -5,17 +5,17 @@
 
 class BlackHole {
 public:
-    BlackHole(double mass, glm::dvec3 position) : m_mass(mass), m_position(position) {
+    BlackHole(float mass, glm::vec3 position) : m_mass(mass), m_position(position) {
         m_radius_s = 2 * Constants::G * m_mass / (Constants::c * Constants::c);
     }
 
-    double GetMass() const { return m_mass; }
-    glm::dvec3 GetPosition() const { return m_position; }
-    double GetRadius() const { return m_radius_s;}
+    float GetMass() const { return m_mass; }
+    glm::vec3 GetPosition() const { return m_position; }
+    float GetRadius() const { return m_radius_s;}
 
 private:
-    double m_mass;
-    double m_radius_s;
-    glm::dvec3 m_position;
+    float m_mass;
+    float m_radius_s;
+    glm::vec3 m_position;
 };
 #endif
